Initial refactoring ofr RFC 2.0.0
