+++ jsconfig.json
@@ -1,6 +1,17 @@
 {
   "compilerOptions": {
-    "checkJs": true
+    "outDir": "dist",
+    "checkJs": true,
+    "allowJs": true
   },
-  "exclude": ["node_modules", "**/node_modules/*"]
+  "exclude": [
+    "node_modules",
+    "**/node_modules/*",
+    "dist",
+    "coverage",
+    "config",
+    ".nyc_output",
+    "test",
+    "util"
+  ]
 }
