+++ lib/InputDeployerVerify.js
@@ -3,7 +3,7 @@ const Jira = require('./Jira')
 const { ENV, ISSUE_LINK_TYPE_NAME, VERIFY_STATUS, REASON } = require('./constants')
 const { previousEnv } = require('./util-functions')
 const RFDWOKFLOW = require('../lib/JiraRfdWorkflowV1.2')
-const RFCWORKFLOW = require('../lib/JiraRfcWorkflowV1.2.1')
+const RFCWORKFLOW = require('../lib/JiraRfcWorkflowV2.0.0')
 
 module.exports = class {
     constructor(settings) {
