+++ lib/Jira.js
@@ -5,7 +5,7 @@
 const JiraClient = require('jira-connector')
 const CONST = require('./constants')
 const RFDWOKFLOW = require('../lib/JiraRfdWorkflowV1.2')
-const RFCWORKFLOW = require('../lib/JiraRfcWorkflowV1.2.1')
+const RFCWORKFLOW = require('../lib/JiraRfcWorkflowV2.0.0')
 
 const JIRA_ENV_REVIEWER = CONST.JIRA_ENV_REVIEWER
 const ISSUE_TYPE_CODE = CONST.ISSUE_TYPE_CODE
