+++ package.json
@@ -17,7 +17,7 @@
       "mocha": "mocha",
       "nyc": "nyc",
       "depcheck": "depcheck",
-      "eslint": "eslint --ignore-path .gitignore .",
+      "eslint": "eslint .",
       "prettier": "prettier  '**/*.js' --check",
       "gen-workflows": "node util/create-jira-workflow-class.js lib/assets/RFC-Workflow-v1.2.1.xml lib/JiraRfcWorkflowV1.2.1.js && node util/create-jira-workflow-class.js lib/assets/RFD-Workflow-V1.2.xml lib/JiraRfdWorkflowV1.2.js && node util/create-jira-workflow-class.js lib/assets/RFC-Workflow-v2.0.0.xml lib/JiraRfcWorkflowV2.0.0.js",
       "tsc": "tsc -p ."
