+++ tsconfig.json
@@ -1,28 +1,1 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist/",
-    "sourceMap": true,
-    "noImplicitAny": true,
-    "module": "commonjs",
-    "target": "ES6",
-    "allowJs": true,
-    "esModuleInterop": true,
-    "strictNullChecks": true,
-    "types": ["node"]
-  },
-  "include": [
-    "./lib/**/*",
-    "./test/**/*",
-    "./util/**/*"
-  ],
-  "exclude": [
-    "./node_modules",
-    "**/*.spec.ts"
-  ],
-  "rules": {
-    "quotemark": [
-      true,
-      "single"
-    ]
-  }
-}
+jsconfig.json
